Migrate tests to plain junit resolving some kotlin-reflect problems
