+++ build.gradle.kts
@@ -10,7 +10,7 @@ val kotlinVersion: String by project
 val kotlinCompilerChecksum: String by project
 val kotlinCompileTestVersion: String by project
 val detektPluginVersion: String by project
-val spekVersion: String by project
+val junitVersion: String by extra
 
 group = "io.github.detekt"
 version = detektPluginVersion
@@ -35,12 +35,9 @@ detekt {
 }
 
 repositories {
-    jcenter()
+    mavenCentral()
     mavenLocal()
     maven { setUrl("https://dl.bintray.com/arturbosch/code-analysis") }
-    maven {
-        setUrl("https://dl.bintray.com/spekframework/spek-dev")
-    }
 }
 
 dependencies {
@@ -53,17 +50,13 @@ dependencies {
     runtimeOnly("io.gitlab.arturbosch.detekt:detekt-core:$detektVersion")
     runtimeOnly("io.gitlab.arturbosch.detekt:detekt-rules:$detektVersion")
     runtimeOnly("io.gitlab.arturbosch.detekt:detekt-formatting:$detektVersion")
-
-    testImplementation(kotlin("reflect", kotlinVersion))
     testImplementation("org.assertj:assertj-core:$assertJVersion")
     testImplementation("com.github.tschuchortdev:kotlin-compile-testing:$kotlinCompileTestVersion")
-    testImplementation("org.spekframework.spek2:spek-dsl-jvm:$spekVersion")
-    testRuntimeOnly("org.spekframework.spek2:spek-runner-junit5:$spekVersion")
+    testImplementation("org.junit.jupiter:junit-jupiter:$junitVersion")
 }
 
 tasks.withType<Test>().configureEach {
     useJUnitPlatform()
-    systemProperty("SPEK_TIMEOUT", 0)
     testLogging {
         events = setOf(
             TestLogEvent.FAILED,
