+++ gradle.properties
@@ -11,7 +11,7 @@ shadowVersion=6.0.0
 # Dependencies
 assertJVersion=3.17.2
 kotlinCompileTestVersion=1.3.0
-spekVersion=2.0.13
+junitVersion=5.7.0
 
 kotlin.code.style=official
 systemProp.sonar.host.url=http://localhost:9000
